New Reflection system * Re-written Reflection system modelled after a dream I had * Some `float`s have been changed to `double`s to work with it * Exists under the `Reflection` namespace * `GenericType` is now `GenericValue`, `PropType` is now `ValueType` * Editor cannot set Vector3's anymore for some reason. Other stuff also probably broke. * HOLY SHIT IT WORKED FIRST TRY WITHOUT CRASHING AFTER 3 HOURS OF REFACTORING OH MY G*D I SLAYED

+++ include/gameobject/Base3D.hpp
@@ -23,10 +23,10 @@ public:
 
 	Vector3 LinearVelocity;
 	Vector3 AngularVelocity;
-	float Mass = 1.0f;
+	double Mass = 1.0f;
 
-	float Transparency = 0.0f;
-	float Reflectivity = 0.0f;
+	double Transparency = 0.0f;
+	double Reflectivity = 0.0f;
 
 	Color ColorRGB = Color(1.0f, 1.0f, 1.0f);
